anpr-service/requirements.txt
# These should already be installed
fastapi
uvicorn
ultralytics
opencv-python
easyocr
numpy
pillow
python-multipart